feat(skills): edit existing Word documents in place via the XML pipeline
The word-documents skill knew one way to produce a DOCX: build it from
scratch with python-docx. So a change to a document the user already had —
their template, an upload, the previous version of the same deliverable —
was answered by regenerating the whole thing. Document() starts from the
library's default template and cannot reproduce a firm's, so the styles,
numbering, headers and footers, and theme fonts came back as generic
defaults and the user got a document that no longer looked like theirs.

Split the skill the same way presentations is split. Nothing existing and no
template: python-docx, unchanged. Anything that starts from an existing
.docx: unpack the package, make targeted edits to word/document.xml and the
other parts, validate, repack. python-docx stays pinned because it is still
the create path.

docx_clean.py backs the check step. Its two checks — every XML part parses,
every relationship target resolves — are the same ones pptx_clean.py makes
and neither depends on the package format, so the shared logic moved into
_openwave_ooxml.py and both scripts are now thin CLIs over it that differ
only in the format they name. Like its counterpart it reports and never
rewrites.

The reopen check splits with the paths: a generated document is still walked
with python-docx, an edited one is proven by LibreOffice loading it during
the existing render step.
