+++ crates/openwave-code-execution/src/lib.rs
@@ -110,9 +110,10 @@ pub const PACKAGE_MANAGER_DOMAINS: &[&str] = &[
 ];
 
 /// Files copied as one indivisible helper library into an exec workspace.
-pub const DOCUMENT_SCRIPT_FILES: [&str; 11] = [
+pub const DOCUMENT_SCRIPT_FILES: [&str; 13] = [
     "_openwave_preview.py",
     "_openwave_calc.py",
+    "_openwave_ooxml.py",
     "render_pdf.py",
     "extract_pdf_figures.py",
     "render_office.py",
@@ -122,6 +123,7 @@ pub const DOCUMENT_SCRIPT_FILES: [&str; 11] = [
     "pptx_clean.py",
     "calc_uno.py",
     "xlsx_recalc.py",
+    "docx_clean.py",
 ];
 
 /// The baseline Python packages guaranteed on every execution backend, as
